Added playwright browser binaries
