--- Uxcheckmate/Uxcheckmate_Main/Services/Abstract/IPa11yService.cs
@@ -1,11 +1,0 @@
-using Uxcheckmate_Main.Models;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-
-namespace Uxcheckmate_Main.Services
-{
-    public interface IPa11yService
-    {
-      Task<ICollection<AccessibilityIssue>> AnalyzeAndSaveAccessibilityReport(Report report);
-    }
-}
