+++ Uxcheckmate/Uxcheckmate_Main/Uxcheckmate_Main.csproj
@@ -21,6 +21,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="9.0.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.2" />
+    <PackageReference Include="Microsoft.Playwright" Version="1.50.0" />
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="9.0.0" />
     <PackageReference Include="PdfSharp-MigraDoc" Version="6.1.1" />
     <PackageReference Include="PdfSharpCore" Version="1.3.65" />
